requirements2.txt
aiofiles==0.8.0
aiosqlite==0.17.0
anyio==3.4.0
appdirs==1.4.3
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asyncua==0.9.92
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.11.1
bleach==4.1.0
blinker==1.4
Brlapi==0.7.0
bs4==0.0.1
certifi==2022.6.15
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==37.0.4
cupshelpers==1.0
dbus-python==1.2.16
debugpy==1.5.1
decorator==5.1.0
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.0.6
django-filter==22.1
django-geojson==3.2.1
djangorestframework==3.13.1
djangorestframework-gis==1.0
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
GDAL==3.0.4
geojson==2.5.0
gpg===1.13.1-unknown
httplib2==0.14.0
idna==3.3
importlib-metadata==1.5.0
importlib-resources==5.4.0
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.0.3
json5==0.9.6
jsonschema==4.2.1
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyter-server==1.13.1
jupyterlab==3.2.5
jupyterlab-pygments==0.1.2
jupyterlab-server==2.9.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lxml==4.7.1
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
monotonic==1.5
more-itertools==4.2.0
nbclassic==0.3.4
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.4
netifaces==0.10.4
notebook==6.4.6
numpy==1.22.0
oauthlib==3.2.0
olefile==0.46
opcua==0.98.13
opcua-client==0.8.3
opcua-modeler==0.5.12
opcua-widgets==0.5.10
opencv-python==4.5.5.62
packaging==21.3
pandocfilters==1.5.0
paramiko==2.6.0
parso==0.8.3
pexpect==4.6.0
pickleshare==0.7.5
Pillow==9.2.0
prometheus-client==0.12.0
prompt-toolkit==3.0.24
protobuf==3.6.1
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pycairo==1.16.2
pycparser==2.21
pycups==1.9.73
Pygments==2.10.0
PyGObject==3.36.0
PyJWT==2.4.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==3.0.6
PyQt5==5.14.1
pyRFC3339==1.1
pyrsistent==0.18.0
PySocks==1.6.8
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python3-openid==3.2.0
pytz==2019.3
pyxattr==0.6.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==22.3.0
reportlab==3.5.34
requests==2.28.1
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
SecretStorage==2.3.1
Send2Trash==1.8.0
simplejson==3.16.0
sip==4.19.21
six==1.14.0
sniffio==1.2.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
systemd-python==234
terminado==0.12.1
testpath==0.5.0
torbrowser-launcher==0.3.2
tornado==6.1
traitlets==5.1.1
typing-extensions==4.0.1
ubuntu-advantage-tools==27.9
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
Unipath==1.1
urllib3==1.26.10
usb-creator==0.3.7
vboxapi==1.0
virtualenv==20.0.17
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
xkit==0.0.0
youtube-dl==2020.3.24
zipp==3.6.0
